feat: Use fn patch version for cli autocomplete
This brings the cli function autocompletion in line with other
updates to kpt function catalog/documentation, in which we display
the fully qualified (patch) version for functions.

This is accomplished using the new /catalog-v2.json endpoint
which serves an updated schema that includes the latest patch
version of each function release.