fix: bash auto completion for CLI flags
This kpt subcommand is invoked by bash shell completion, which
previously returned an error when attempting to perform
auto completion for flag values. The actual subcommands and their
associated flags are registered with the root Command but not
this subcommand, which results in flag validation returning an
error for this subcommand.

Disabling flag validation fixes this issue and enables auto
completion for flags in bash.