refactor: set klog flags before adding to cmd
When flags are set using the Command.Flags().Set() method, the flags
are marked as Changed. This results in the Cobra autocompletion logic
treating it as though the user passed flags for the Command, ultimately
resulting in Cobra not rendering the available subcommands for the root
kpt command.

This change enables the expected behavior when invoking `kpt <tab>` with
the Cobra built-in autocompletion.