refactor: use Cobra built-in autocompletion
Prior to this change, the kpt CLI exposed two methods for enabling shell
autocompletion:
- kpt --install-completion
- kpt completion [bash|fish|zsh|powershell]

The former was implemented using complete, whereas the latter is
provided by Cobra as a default. There is not a clear motivation to
offer two methods to provide autocompletion functionality, so this
change aims to consolidate these into a single method.

This change removes the --install-completion option and applies some
fixes to enable proper operation of Cobras autocompletion. The decision
was made to favor the Cobra implementation over the custom one in order
to reduce maintenance cost and offer similar behavior to other popular
CLIs which leverage Cobra (e.g. kubectl)