Cobra based autocompletion (#2631)
* refactor: set klog flags before adding to cmd

When flags are set using the Command.Flags().Set() method, the flags
are marked as Changed. This results in the Cobra autocompletion logic
treating it as though the user passed flags for the Command, ultimately
resulting in Cobra not rendering the available subcommands for the root
kpt command.

This change enables the expected behavior when invoking `kpt <tab>` with
the Cobra built-in autocompletion.

* refactor: use Cobra built-in autocompletion

Prior to this change, the kpt CLI exposed two methods for enabling shell
autocompletion:
- kpt --install-completion
- kpt completion [bash|fish|zsh|powershell]

The former was implemented using complete, whereas the latter is
provided by Cobra as a default. There is not a clear motivation to
offer two methods to provide autocompletion functionality, so this
change aims to consolidate these into a single method.

This change removes the --install-completion option and applies some
fixes to enable proper operation of Cobras autocompletion. The decision
was made to favor the Cobra implementation over the custom one in order
to reduce maintenance cost and offer similar behavior to other popular
CLIs which leverage Cobra (e.g. kubectl)

* ci: allow no mozilla repos in create-licenses

As of a prior commit there are no dependencies on mozilla repos.

This change prevents the script from failing in the case of
mozilla_repos being an empty string.