feat: always print truncated function stderr (#2639)
* feat: always print truncated function stderr

This change makes it so that kpt fn render|eval will always print
truncated stderr output from functions. This is intended to improve the
experience of debugging functions.

* test: add unit test cases for printFnStderr

* test: add e2e test for fn eval stderr

This test case is intended to check that kpt fn eval outputs the stderr
of the function even when the function succeeds. This is doing using a
simple function script which emits hello world to stderr.

* test: add e2e test for fn render stderr

This test case is intended to check that kpt fn render outputs the stderr
of the function even when the function succeeds. This is doing using a
simple function script which emits hello world to stderr.