Delete PackageRevisions in Git
If package has its own Git ref (own branch or package-specific tag),
the ref is deleted. This is true for drafts, proposals, and final
packages tagged with their own tag (refs/tags/<package patch>/revision).

If the package is tagged with general tag of a different format,
we can't delete it and deletion fails.

If the package is in a branch, we delete it by creating a commit
that removes it from that branch.

This change also includes few bug fixes that ensure that we track
package lifecycle correctly through package updates.