Respect Registration Directory (#3017)
If a sub-directory is specified at repository registration, the repository will
only include package revisions in the specified directory. Likewise, it will
only create packages in the specified subdirectory.

In order to maintain stable resource names within the repo (regardless of the
registration parameters), the package names are unaffected by the directory
projection.