fix: Fix broken releases due to improper configs
Add release as a phony to our Makefile and remove unnecessary steps in
the release workflow. Certain elements from the previous iteration of
the workflow should also be removed since GoReleaser is now being run
within a container and not as a native GitHub action, so they are not
compatible.