Propagate ResourceGroup labels and annotations
When I declare a ResourceGroup on disk, I expect the labels and
annotations defined on it to be propagated to the in-cluster inventory.
That was previously not the case.

kioutil annotations are removed (path, index, etc) since they are
redundant for inventories that are typically defined in one place, and
don't need to be reconstructed by a function pipeline.

Signed-off-by: Fredrik Sommar <fred.sommar+github@gmail.com>